Add Controllers, edit models, add specification, edit repository

+++ WebAutopark.Core/Entities/VehicleType.cs
@@ -6,6 +6,6 @@ namespace WebAutopark.Core.Entities
     {
         public string TypeName { get; set; }
 
-        public decimal TaxCoefficient { get; set; }
+        public double TaxCoefficient { get; set; }
     }
 }
